+++ changelog.md
@@ -1,3 +1,6 @@
+- feat(sh): :fire: #45 Compatible with AdGuard
+- 1. Compatible with AdGuard, Skip "AdGuard Intermediate CA"
+---
 - chore(UI): :tada: Adapt to dark mode
 ---
 - Merge pull request #41 from SafaSafari/iyue